pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pianocli"
version = "1.0.0"
authors = [
  { name="Manasvi Bathula", email="manasvi.bathula@gmail.com" },
]
description = "A playable piano in your terminal."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
# --- List of libraries your app needs to run ---
dependencies = [
    "textual",
    "pyfluidsynth",
    "typer[all]", 
]

[project.scripts]
termpiano = "pianocli.main:cli"